pyproject.toml
[tool.poetry]
name = "sideproject3"
version = "0.1.0"
description = ""
authors = ["Kk <iankk0806@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
